flowshield-udrl/requirements.txt
# FlowShield-UDRL Dependencies
# Last updated: January 2026

# ============================================
# Core ML Framework
# ============================================
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# ============================================
# Reinforcement Learning
# ============================================
gymnasium>=0.29.0
stable-baselines3>=2.0.0

# ============================================
# ODE Solvers (for Flow Matching)
# ============================================
torchdyn>=1.0.0

# ============================================
# Logging & Progress
# ============================================
tensorboard>=2.13.0
tqdm>=4.65.0

# ============================================
# Visualization
# ============================================
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=9.0.0

# ============================================
# Data Processing
# ============================================
pandas>=2.0.0

# ============================================
# Development (Optional)
# ============================================
pytest>=7.3.0
black>=23.0.0
isort>=5.12.0

# ============================================
# Documentation (Optional)
# ============================================
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
